requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djongo==1.3.6
pylint==2.16.3
pylint-django==2.5.3
sqlparse==0.4.4
pymongo==3.12.1
django-filter==24.3